Take a huge step forward in overcoming problems with virtual function types.  Essentially, the callbacks being passed from C# to C++ were not dealing with the return types correctly.  Any delegate that returned a reference type would cause both Mono and the CLR to crash, thus indicating quite clearly that this was not an option.  To overcome this, the C# class template now generates adapters for virtual functions that hide the return types.  The adapter is passed as the delegate rather than the virtual function.
Among other changes in this commit, the C# class template now exports
custom marshalers for abstract classes.  This brings up a rather big issue
relating to passing pointers to abstract types because they cannot be
marshaled into a corresponding C# type through any means I can think of
(so far).  This requires further pondering.

With this change, I am able to run a simple test application that
instantiates a vpr.GUID in C# and passes it over to native C++ code for
handling.  Trying to do polymorphic handling of a vpr.GUID as a
vpr::SerializableObject* is not working, however, due to the aforementioned
problems with pointers to abstract types being passed to C#.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@140 f34b864f-2b4e-4307-96f1-59446bce9074