Bug fixed:	A pointer to an object of an abstract types could not be 		marshaled from C++ into C#.
In a custom marshaler for an abstract type, generate a dummy class that
implements the abstract methods of the abstract type in question.  This
allows the custom marshler for the abstract type to return something when
a native object comes into C# land.  That in turn allows methods that
take instances of the abstract type to get something they can work with,
even when the real implmentation type is not known at compile time.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@142 f34b864f-2b4e-4307-96f1-59446bce9074