Initial pass at supporting proper C++ object instantiation when abstract classes are involved.  For a concrete C# class that implements the interface defined by an abstract base class, we need the abstract base class to allocate a C++ object (a bridge class instance) so that we have something meaningful to pass to the native code.
To accomplish this, the code for allocating the delegates has been
moved into a new private method (redeclared for each class, which
appears to be valid for C#) called allocDelegates().  This simplifies
the contents of all the generated constructors.  Next, a special
protected constructor is generated for each abstract class that takes
a class-specific object of type NoInitTag.  In this constructor, only
allocDelegates() is called.  In other constructors, mRawObject is
initialized just as before, but in the case of the other constructors,
the specialbase class constructor will be called to prevent mRawObject
from being initialized.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@162 f34b864f-2b4e-4307-96f1-59446bce9074