Refactor:	Users now define the name of the module in the input 		files they write rather than naming the module on the 		command line.  This allows multiple modules to be exposed 		to the CLI from a single input file or from multiple input 		files provided during a single run of Sharppy.
My intention with this is to move towards using Sharppy to expose many
things to the CLI at once so that proper inheritance information can be
determined automatically.  Running the tool with only a single input file
targeting a single "module" does not provide all the information we need.
For example, if a class in one module inherits from a class in another,
that information is not available to Sharppy unless both classes are
handled together.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@169 f34b864f-2b4e-4307-96f1-59446bce9074