Simplify the decision about what type to use for the "self" parameter passed to C wrapper functions.  We now reuse $class_name, which is determined based on whether a bridge class exists to handle virtual methods.
The main reason for changing this is to allow virtual methods that would
otherwise be inaccessible (protected) to be called by the C wrapper
functions.  I suppose that the C wrapper function could be declared as a
friend of the C++ bridge class, and that may still need to happen.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@187 f34b864f-2b4e-4307-96f1-59446bce9074