This is an attempt to improve the names of methods used for querying the names/identifiers of types in declarations.Declaration.  With this change, I have attempted to reduce the number of name accessor methods and to make the naming convention more consistent.
Buried in all of this is a bug fix for some templated classes not having
their names made friendly to C# when they types appeared as function
parameters that required marshaling.  This is what I was trying to fix in
the first place, but I finally got fed up with the method naming in
declarations.Declaration.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@227 f34b864f-2b4e-4307-96f1-59446bce9074