Refactor:	Make the read-only property RawObject public.  Remove the 		internal visibility of mRawObject. Bug fixed:	Generated code could fail to compile with the Microsoft C# 		compiler due to visibility issues with external assemblies 		trying to access mRawObject. RF Bug#:	17
I am not entirely happy with making the property RawObject public, but
I cannot come up with anything better.  Originally, I had expected that
the custom marshalers would prevent any need for mRawObject (or RawObject)
to be visible to outside code.  This is the second time that I have had to
make significant changes to the visibility of the raw pointer value for
native object mapping.  The first time was the result of poor planning on
my part alone.  This time, the problem could have been identified right
away if Mono's mcs tool weren't so buggy.


git-svn-id: file:///Users/patrick/nobackup/aeon-svn/trunk/sharppy@345 f34b864f-2b4e-4307-96f1-59446bce9074